Create WIN_get_SQL_cluster_info.ps1
Because our SQL instances are on an MS CLUSTER, and the unknown was which instance was running on which physical cluster, this script performs a query on the physical servers in the CLUSTER and saves the information about instances running on the cluster into JAMS variables.  Other programs read these variables, and are able to dynamically backup the SQL instances on each physical server .  (We use TSM TDP for SQL).