Create LOG_SUMMARY.ps1
We have multiple jobs that run, and wanted a way to collect all of the logs into one document.  This script does that.. Each job that runs, creates a JAMS variable that contains the JAMS Entry + JAMS LogFileName.  This script reads these variables, reads each logfile, and outputs one logfile with a banner inbetween each of the individual log files.  The JAMS JOB that runs this, has the email notification set with a minimum completion code of SUCCESS, and an email with this log as an attachment is sent to our Sharpoint server.  The Sharepoint server in turn stores this into an audit record archive, where Sharepoint automatically handles the storage of these audits, aging, etc...  This creates a seamless reboot process from start to finish for us.